Move tallyVotes function for easier testing
